pyscripts/requirements.txt
amqp==5.1.1
asgiref==3.7.2
autopep8==2.0.4
billiard==4.1.0
celery==5.3.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.7
Django==4.2.6
django-celery-beat==2.5.0
django-dbsettings==1.3.0
django-filter==23.3
django-grappelli==3.0.8
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
kombu==5.3.2
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pika==1.3.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyactiveresource==2.2.2
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
ShopifyAPI==12.3.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
